+++ CHANGELOG.md
@@ -1,6 +1,28 @@
 # Changelog
 
-## v[v1.2.1] - 2023-10-11
+## ?? - 2023-10-31
+
+### Added
+
+- `DetectAndNotifyInactiveUsers` detects inactive users and sends a reminder message to them to login again.
+- `CleanupUsersMarkedForDeletion` deletes all user accounts that do not react after a certain time. All user tokens are removed and study-service is notified. Users are informed by message about the deletion of their account.
+
+New environment variables:
+- `NOTIFY_INACTIVE_USERS_AFTER`: time after which inactivity notification will be triggered in seconds.
+- `DELETE_ACCOUNT_AFTER_NOTIFYING_USER`: length of interval between user notification and deletion of user account in seconds if user does not react. 
+
+Both variables must be defined and greater than zero to activate deletion workflow. If workflow is active the following must be provided:
+
+Email templates:
+- `account-inactivity`: invites inactive user to login to account in order to prevent account deletion,
+- `account-deleted-after-inactivity`: informs user that account is deleted.
+
+Environment variable:
+- `ADDR_STUDY_SERVICE`: address of study service.
+
+
+
+## [v1.2.1] - 2023-10-11
 
 ### Changed